+++ tests/test_project/inherited_flatness.py
@@ -1,0 +1,6 @@
+from machinable import Component
+
+
+class InheritedFlatness(Component):
+
+    pass